test fixed and moved to kernel_tests/rnn_cell_test.py
